feat: move genesis to keeper level
