+++ x/fundraising/keeper/genesis_test.go
@@ -1,20 +1,25 @@
-package fundraising_test
+package keeper_test
 
 import (
 	"time"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	tmproto "github.com/tendermint/tendermint/proto/tendermint/types"
 
-	"github.com/tendermint/fundraising/app"
-	"github.com/tendermint/fundraising/testutil/simapp"
 	"github.com/tendermint/fundraising/x/fundraising"
 	"github.com/tendermint/fundraising/x/fundraising/types"
 
 	_ "github.com/stretchr/testify/suite"
 )
 
-func (s *ModuleTestSuite) TestGenesisState() {
+func (s *KeeperTestSuite) TestDefaultGenesis() {
+	genState := types.DefaultGenesisState()
+
+	s.keeper.InitGenesis(s.ctx, *genState)
+	got := s.keeper.ExportGenesis(s.ctx)
+	s.Require().Equal(genState, got)
+}
+
+func (s *KeeperTestSuite) TestGenesisState() {
 	auction := s.createFixedPriceAuction(
 		s.addr(0),
 		sdk.OneDec(),
@@ -45,10 +50,10 @@ func (s *ModuleTestSuite) TestGenesisState() {
 	s.Require().Equal(types.AuctionStatusStarted, auction.GetStatus())
 
 	// Place bids
-	s.placeBid(auction.GetId(), s.addr(1), types.BidTypeFixedPrice, sdk.OneDec(), parseCoin("20000000denom2"), true)
-	s.placeBid(auction.GetId(), s.addr(2), types.BidTypeFixedPrice, sdk.OneDec(), parseCoin("30000000denom2"), true)
-	s.placeBid(auction.GetId(), s.addr(3), types.BidTypeFixedPrice, sdk.OneDec(), parseCoin("15000000denom2"), true)
-	s.placeBid(auction.GetId(), s.addr(4), types.BidTypeFixedPrice, sdk.OneDec(), parseCoin("35000000denom2"), true)
+	s.placeBidFixedPrice(auction.GetId(), s.addr(1), sdk.OneDec(), parseCoin("20000000denom2"), true)
+	s.placeBidFixedPrice(auction.GetId(), s.addr(2), sdk.OneDec(), parseCoin("30000000denom2"), true)
+	s.placeBidFixedPrice(auction.GetId(), s.addr(3), sdk.OneDec(), parseCoin("15000000denom2"), true)
+	s.placeBidFixedPrice(auction.GetId(), s.addr(4), sdk.OneDec(), parseCoin("35000000denom2"), true)
 
 	// Modify the current block time a day after the end time
 	s.ctx = s.ctx.WithBlockTime(auction.GetEndTimes()[0].AddDate(0, 0, 1))
@@ -71,30 +76,12 @@ func (s *ModuleTestSuite) TestGenesisState() {
 
 	var genState *types.GenesisState
 	s.Require().NotPanics(func() {
-		genState = fundraising.ExportGenesis(s.ctx, s.keeper)
+		genState = s.keeper.ExportGenesis(s.ctx)
 	})
 	s.Require().NoError(genState.Validate())
 
 	s.Require().NotPanics(func() {
-		fundraising.InitGenesis(s.ctx, s.keeper, *genState)
+		s.keeper.InitGenesis(s.ctx, *genState)
 	})
-	s.Require().Equal(genState, fundraising.ExportGenesis(s.ctx, s.keeper))
-}
-
-func (s *ModuleTestSuite) TestMarshalUnmarshalDefaultGenesis() {
-	genState := fundraising.ExportGenesis(s.ctx, s.keeper)
-	bz, err := s.app.AppCodec().MarshalJSON(genState)
-	s.Require().NoError(err)
-
-	genState2 := types.GenesisState{}
-	err = s.app.AppCodec().UnmarshalJSON(bz, &genState2)
-	s.Require().NoError(err)
-	s.Require().Equal(*genState, genState2)
-
-	app2 := simapp.New(app.DefaultNodeHome)
-	ctx2 := app2.BaseApp.NewContext(false, tmproto.Header{})
-	fundraising.InitGenesis(ctx2, app2.FundraisingKeeper, genState2)
-
-	genState3 := fundraising.ExportGenesis(ctx2, app2.FundraisingKeeper)
-	s.Require().Equal(genState2, *genState3)
+	s.Require().Equal(genState, s.keeper.ExportGenesis(s.ctx))
 }
